Refactor network proxy      - Refactor cmd/proxy/main.go into a library under pkg/proxy.      - Add logger package for logging functionality.      - Rename certfmt to cert to contain generalized certificate functions.      - Move cert and docker packages to pkg.
